% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame of numeric and string values intended as training data (with no NA values).}

\item{cl}{The true class value of the training data, \code{train} (with no NA values).}

\item{k_nn}{Numeric value representing the number of neighbors.}

\item{k_cv}{Numeric value representing the number of folds for cross-validation.}
}
\value{
List containing predicted class (y-hat) for all observations in \code{train}, and a numeric
value representing the cross-validation misclassification error.
}
\description{
Predicts an output class of a given training data using the k-Nearest Neighbors non-parametric
statistical prediction method and cross-validation.
}
\examples{
ex_penguins <- na.omit(my_penguins)
my_knn_cv(train = ex_penguins[,c(3:6)], cl = ex_penguins$species, k_nn = 1, k_cv = 5)

}
\keyword{prediction}
