% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Input of formula class.}

\item{data}{Data frame of numeric values.}
}
\value{
Table where coefficients are represented by a row, and columns are the estimate (numeric),
standard error (numeric), t-value (numeric), and the p-value of the t-test (numeric).
}
\description{
Fits a linear model to a data frame.
}
\examples{
my_lm(formula = x ~ y, data = mydata)
my_lm(formula = x + z ~ y, data = dataset)

}
\keyword{inference}
