% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric value representing the number of folds for cross-validation.}
}
\value{
Numeric value representing the cross-validation error.
}
\description{
Predicts an output class of a given training data using the Random Forest
statistical prediction method and cross-validation.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
