% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-Test Function}
\usage{
my_t.test(x, alternative = c("two-sided", "less", "greater"), mu)
}
\arguments{
\item{x}{Vector of numeric data (with no NA values) that t-test will be performed on.}

\item{alternative}{String input of either: "greater", "less", or "two-sided".}

\item{mu}{Numeric value that represents the true mean of the input data, \code{x}.}
}
\value{
List containing the test statistic (numeric), degrees of freedom (numeric),
initial input value of \code{alternative}, and the p-value (numeric).
}
\description{
Performs one or two-sided t-tests on numerical data.
}
\examples{
my_t.test(x = data[1:100], alternative = "less", mu = 0)
my_t.test(x = mydata[5:200], alternative = "two-sided", mu = 1)

}
\keyword{inference}
